requirements.txt
certifi==2021.5.30
chardet==4.0.0
cssselect==1.1.0
idna==2.10
lxml==4.6.3
numpy==1.20.3
pandas==1.2.4
pyquery==1.4.3
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
urllib3==1.26.5
